requirements.txt
aiohttp==3.9.1
async-lru==2.0.4
asyncpg==0.29.0
beautifulsoup4==4.12.2
discord.py==2.3.2
fake-useragent==1.4.0
Pillow==10.1.0
python-dotenv==1.0.0
python-steam-api==1.2.0
regex==2023.10.3
tweepy==4.14.0
twitchAPI==4.1.0
validators==0.22.0